Provide Legal SIN and Licenses
This enhancement supplies a Custom Data Dictionary to add a legal SIN and licenses to the Gear tab. There are restrictions:
1) The character must have one of the SINner qualities.
A legal SIN is provided by a government or corporation, so the character must be in the system.
2) Only legal licenses can be added to a Legal SIN.
After all why would a character spend large amounts of Nuyen to add a Fake License to a real SIN?
Even a rating 6 fake can be spotted as a fake; resulting in court appearances, fines, jail and maybe the SIN ending up as a criminal SIN.
3) The cost of a legal license is variable, so the gamemaster will have to decide the cost using real world examples as guidelines.