Fixed MCT/WInchester-Minigrenade (SL
Since the grenade was already partially created, I had problems inserting it correctly (grenades have to be created twice, which is why I did not understand at first, it was not visible). When I finally had the expected result in the folder of the last Chummer-nightly, I simply forgot to move it to the Github folder.

Sorry about that. Now it should hopefully work as expected.