Adjusted BMW M8 (#3397)
The vehicle 'BMW M8' is so far listed with 'Source: SHB'. I do not have this source book. However, the vehicle is also included in the German Rigger 5.0  ('R5G') and has almost the same values there. Only the acceleration is slightly higher with '3' instead of '2'.

Since the edition of Rigger 5.0 (in German 'Asphaltkrieger') is from the year 2016, this change could be considered as a correction.

So I suggest to change the source of 'BMW M8' from 'SHB' to 'R5G' (including new page numbers) and adjust the acceleration. Since the vast majority of German users probably owned R5 instead of SHB as their source book, I think this might be a good solution.

If I have missed something in the proposal or if I have made something worse: please let me know or reject the suggestion.