Datapuls Sota 2080 (#3499)
* Datapuls Sota 2080

Added German book 'Datapuls Sota 2080' (SOTG), mentors from SOTG, one heavy pistol and some melee weapons

* Update weapons.xml

* Erased addition of 'Sota 2080'

I am still not sure, if new folding spade of 'Datapuls SOTA 2080' should be treated similar to grenades (only available in weapons but also shows as gear).

But as I can not use 'addgear' in weapons.xml, I decided to leave 'folding spade' of SOTA 2080 in wepaons.xml and 'collapsible shovel' of Run Faster in gear.xml (as it is written in the books). Both is translated in German as 'Klappspaten' - one is gear (RF) and the other one is a weapon (SOTG).