Addition to custom rule for German Yamaha Kaburaya (#3975)
Besides the acceleration, the costs for the motorbike 'Yamaha Kaburaya' are also different according to German rules. Since this is mentioned not only once but twice and is not mentioned in any errata thread in German forums, this adjustment seems to be intentional. The custom rule has been modified to include both.

Licence text has been added.