Removed blnConcatSelectedValue from improvements.
Looking over places where CreateImprovements is used, this parameter is
either:
1. set to false
2. set to true, but corresponding improvement removal code treats it as
though it is set to false.

Initially I opted for converting it to false everywhere, but
AddImprovementCollection contained a lot of messy and inconsistent code
around that parameter, so I decided to clean up.

RemoveImprovements also got a compatibility fix that tries to find stuff
affected by relevant bugs previously.

Fixes #4088.