Fixed a translation error in the German Data and added two German data changes
The spell "turn to goo" was translatet to "Schleimverwandlung" but is actually called "Versteinerung".

The quality "Geschickter Verteidiger" has a Karma cost of 3 in the German version, which was added to the "German Data Changes - Quality Karma Costs".

The spell "turn to goo" has a DV of -2 instead of +3 in thge german version. This was introduced in the new "German Data Changes - Spell Parameter Changes" custom data.