Fixed "trigger removal" applying it's bonus twice
Because the stat blocks in the rulebooks always assume all standard accessories are equip the actual precision of weapons with trigger removal is one lower.

This was already done for some weapons, e.g. the HK Urban Striker, but not for the Walther P118 (RG) and the Cavalier Arms Falchion (GH3).