Removed "Trigger Removal" applying twice, improved some translations, added german data changes
Rulebook statblocks assume, that all accessories are installed. Therefore Chummer applied trigger removal twice. This was fixed by reducing the stated accuracy of all guns affected guns by one.

The optional rule "AlternateMetatypeAttrubuteKarma" description was completely unreadable in German and now accurately states it's function.

The cost of "Perceptive Defender" is 3 instead of 4 in the German version of run and gun. This was added to the "German Data Changes - Qualities".

An alternate rule interpretation for the in debt quality from the German rules was interpreted as German Data Change.