Changes to character sheets
There were several sets of changes loaded to pull #4308 but only the last set were added to Chummer. This pull applies the missing changes.

**Character sheets css definitions file**
The main update in the original pull was standardization of the css definitions in the character sheets.

Several years ago there was a request to prevent column heading being in uppercase for German versions of the Chummer 5* character sheets. This was done by moving the definitions to a separate xslt file for each language folder, without uppercase for the German and Portuguese versions. Since then several changes were made in the base version (used just for the US sheets) that were not applied to the separate language versions.

My original intention was to use a single version of the css definitions file for all languages except for German and Portuguese, minimizing the number of css files that needed to be maintained. Delnar Ersike indicated that template overrides could be used to change css definitions. So I added overrides for German and Portuguese to prevent uppercase folding meaning only one css definitions file is now needed.

In addition Delnar suggested the css file could be used in the other character sheets, as many of the sheets have the same values. I renamed the "Shadowrun 5 set CSS.xslt" file as "xs.Chummer5CSS.xslt" to show it is for general use not just the Shadowrun 5* sheets. Some of the character sheets override default values such as font size.

The common definitions file is not used by the Fancy Blocks and Vehicles sheets because there is so little overlap of the css definitions.

**Calender sheet**
Added use of common css file definition.

**Commlinks sheet**
Added use of common css file definition.

**Contacts sheet**
1) Added use of common css file definition.
2) Correctly set column headings formats for Contact archetype information.

**Dossier sheet**
Added use of common css file definition.

**Expenses sheet**
1) Added use of common css file definition.
2) Adjusted positioning of "Total" to better fit the column

**Fancy Blocks sheet**
Only show "Current Form" box for a shapeshifter when a shifted form is being displayed.

**Formatted Text-Only sheet**
1) Added use of common css file definition.
2) Only show "Current Form" box for a shapeshifter when a shifted form is being displayed.
3) Split skill name line if the length of name pushes the values out of position.
4) Corrected formatting of Attributes spacing when enhanced values are present.
5) Correct expenses amount columns to line up values.
6) Adjusted double space between literal and value for various fields.
7) Print appropriate literal for Tradition or Stream depending on the type of character.
8) In various places the "Rating" literal and value were printed without a space. Corrected.

**Game Master Summary sheet**
1) Added use of common css file definition.
2) Correct display of Martial Arts technique to just the name instead of the entire node.

**Notes sheet**
1) Added use of common css file definition.
2) Reviewed character file to identify as many notes nodes as possible.
Added sections to the Notes sheet to display these notes:

Powers
Martial Arts and any maneuvers
Cyber/bioware
Armor
Melee and Ranged Weapons
Gear
Devices
Programs

Thanks to Delnar Ersike for determining how to display notes from nested items beyond immediate children (a gear item can have sub-items that can also have sub-items, any of which can have notes).

**Shadowrun 5 character sheets**
1) Added use of common css file definition.
2) Modified printing of notes for martial Arts and maneuvers to match Notes sheet.
3) Modified the formatting of Ranged Weapons table to show the quantity of thrown weapons (from Gear) instead of the Ammo fixed amount of 1 (issue #3861).
4) Only show "Current Form" box for a shapeshifter when a shifted form is being displayed.
5) Removed display of notes from the Gear columns - even small notes can mess with the layout of the columns. These notes have been added to the Notes sheet.
Note: This includes notes from devices, that are already displayed in the Devices section.
6) Modified display of Armor list to separate Equipped and other into separate blocks.
7) Sort the list of weapons by Location (similar to the way gear is displayed).

**Sprites and Spirts sheet**
Added use of common css file definition.

**Text-Only character sheet**
1) Added use of common css file definition.
2) Split skill name line if the length of name pushes the values out of position.
3) Sheet originally printed the Base value of skill but missed the Karma amount. Changed the literal to "Rtg" and added the Base and karma values together.
4) In various places the "Rating" literal and value were printed without a space. Corrected.

**Vehicles sheet**
The formatting of the Ranged Weapons section to show the quantity instead of Ammo instead of a fixed amount of 1 has been added automatically as the sheet uses the common ranged weapons formatting file.

**other changes**
xs.fnx.xslt - modified to correctly process leading zeroes and decimals.

Created xt.MartialArts.xslt to format notes from Martial Arts and any maneuvers. This is used by the Notes and Shadowrun 5* sheets.

xs.RangedWeapons.xslt - this is where the changes to the display of Thrown Weapons have been made. This is used by the Shadowrun 5* and Vehicles sheets.