Introduced SustainedSpell class
Quick introduction of the SustainedSpell class.

For quick implementation I stole a bunch of code from spells.cs to load the different display names, since the names not only should be but have to be identical and should also be loaded from the spell.xml this should work without further problems. (Correct me if I'm wrong I honestly can't pretend to understand everything right now).

The base functionality of adding, removing, saving and loading seem to work on first glance.