Cut down on unnecessary declarations. Interface improvements
Loading and Saving, as well as printing now all use interfaced methods, this way they can be saved and loaded in the same foreach instead of needing their own.

The load in clsCharacter uses a switch to differentiate between the objects according to an type node.

Maybe some way to move the .Load and .Add  out of the switch could be found.