Added "Große Mutter (SG)" as custom German Mentor removed redundant translation
The German Street Grim has the mentor "Große Mutter" (Great Mother). It is nearly identical to the one that is present in SASS, but not completely.

Since it is functionally a bit different and it's very unclear if they are meant as the same, adding it as another custom Mentor was the easiest way to include it, without having to make a rules call.