Implemented Save/Load/Print derived
Removed Load() and Print() from ISustainable.
Load(): was unnecessary in the first place.
Spell.Print(): Has a different signature to all other Prints(), changing this would have been a bigger hassle than handling it as a switch.

Modfied Create() to be CreateByMapping(): This copies over all property values from an base object to the derived object. Most of those aren't needed, but can't hurt to have them.

Removed a duplicate line in CharacterOptions.cs
Added an if in Skill.core.cs that got lost somehow.