First basis for the expansion
- Implemented all needed Properties.
- Implemented a Create() Method wich is used to load and set all the necessary information from the manifest.xml.
- Implemented a first version of an Save() Method to write full manifest files (untested).
- Implemented a first version of information display in the CharacterOptions (still quite ugly).
- Implemented a first version of an frmManageDependencies, which will allow to manage dependencies and exclusivities (not fully functional).
   -> Will be reworked to an full on Manifest.xml creator.
- Implemented an SHA512 based Hash system, which creates an hash from all files in the CustomDataDirectory.
   -> Excludes the Manifest.xml to not change the hash as soon, as we modify anything.
- Implemented Methods to check if the correct dependencies are loaded or if an exclusivity is loaded.

Planned:
- The frmSelectDependencies should be reworked to allow the full on creation and editing of the manifest.xml.
- Find a solution to display long text, that don't have escape sequences in an acceptable way.

Important for Review:
I moved to Methods up in the program.cs. They both need to be called befor any windows are created. I don't create any windows in my Create() Method, but it still crashes Chummer if this is included in the Constructor. (Could it be the XML Manager?)
CustomDataDirectories also can't be done with lazy<>, because the created object is added to an list right away. And I really don't want to start refactoring this list and everything behind it.
My previous Workaround was to call  If(!isCreated)Create() in any get accessor. This works well, until it leads to stack overflows and other problems, when not paying attention while changing stuff. (It's also ugly)