Fixed a Bug with Multiple Characters with Different Skill Datasets
If multiple characters were loaded where some characters had different skills defined in their datasets, the caching system for whether a certain skill category was for knowledge skills or not would break and could cause certain skill categories to be flagged as the wrong thing. Fixed this by removing the caching altogether as well as the need for it in the first place (it seems like it was only really necessary for debugging legacy loading of skills from very old Chummer save files).