Optimized Code for Checking Against Item Counts
When checking if a character has too much of a certain thing, Chummer no longer iterates through every single item all the time if it is not meant to show a message of how many of the particular item there actually are. Instead, if the limit is exceeded in the middle of iterating through items, the code will just exist out of counting immediately. In addition, if the code is meant to extend limits to also include additional items, it will not iterate through the same list twice; everything will be processed in a single set of iterations.