Added Ability to Fetch Used Improvements in Value, Removed Boilerplate
First part means that methods that get the values of improvements can now also fetch the improvements that were used to get that value in the first place, which can be used to fix inconsistent tooltips and other UI inconsistencies.
Boilerplate in this case means duplicated code across ImprovementManager.ValueOf() and ImprovementManager.AugmentedValueOf(). The functions now both use a common, private, meta function that takes in the cached dictionary to use for storing values and the function to use to get values from improvements as extra arguments.