Groundwork for the 4e Critter Custom Data File
Laying out the groundwork for the 4e Critter Custom Data File. The First Step is to remove ALL of the 4e critters from the ungodly mess that is critters.xml and while I'm at it sort the damned thing so it is less of an ungodly mess. It's still an ungodly mess, but slightly less so now.

All 4e critters are now moved to a totally non-functional amends file. This will serve as the groundwork for a larger project to slowly start converting the critters to be functional in Chummer5 if someone wishes to enable the amends and use the critters.