Update gear.xml
This changes the cost of the Internal Synth from No Future to a cost equal to the parent cost, effectively doubling the cost of the instrument, rather than tripling it. 

The basis for this change is consistency with the internal Smartgun system's cost. The table for the internal smartgun also lists the cost as {Cost of weapon} x2. From that alone one might think that the internal smartgun system therefore costs twice as much as the weapon cost. However, the text below explains that what this actually means is it "doubles the cost of the weapon"