First Step to Fixing AsyncFriendlyReaderWriterLock's Remaining Problem
Description of the problem:

"This crash happens when enumeration over one of Chummer5a's locking collections that's supposed to maintain thread-safety throws an exception because the collection has changed during the enumeration. Creating the enumerator is supposed to acquire a reader lock on the collection, and acquiring a reader lock is supposed to prevent writer locks from being acquired until it is released. This would normally create issues around reentrancy though (i.e. trying to acquire a writer lock while inside a code block that already acquired a reader lock, which would create a deadlock), so in this and past versions of Chummer5a, the writer lock will only also try to acquire the reader lock if no other reader locks and no writer locks are set. This 'fix' is also where the problem lies: what happens when a reader lock gets acquired first, and then an independent writer lock gets acquired. When the reader lock is acquired, no writer locks are set, so it sets the reader lock semaphore as necessary and breezes through. When the writer lock is acquired though, it first sets the writer semaphore (as it should) and then sees that there is already an active reader. What it should do is realize that the active reader is not one inside of which it is nested, meaning it should wait for that active reader to exit before continuing. It doesn't do that and just continues forward, resulting in it potentially writing the object while the reader lock is still active, which can then result in a situation that causes the aforementioned crash (enumeration crashes because the collection has changed while it was reader-locked)."