Made DebuggableSemaphoreSlim Not Always Track Holders in Debug Mode
When the define to track the last holder of a DebuggableSemaphoreSlim is set, instead of all semaphores being tracked, the developer must now specify in the semaphore's constructor whether they want it tracked or not. This should make the program generally more usable in this mode, since otherwise the massive amount of semaphores used in the program's many reader-writer locks would grind performance to a halt.