Implemented Reduced-Allocation Versions of XmlNode Content Getters
Normally, properties like InnerText, InnerXml, and OuterXml for XmlNodes and XPathNavigators require calling a string builder to build the output string each time. This creates a lot of GC pressure on top of being potentially inefficient in cases where we're only interested in some small part of the output, so I've added extension methods that should help with that by reducing the amount of string builder allocations we need.