Mass Unboxing of String Concatenation
In a lot of cases, we concatenate a string with a constant (e.g., a char), which could end up potentially slow things down because of the required ToString() call to turn that constant into a string whose value we'd already know at compile time. To make sure this doesn't happen, all cases where we would concatenate with compile-time constant non-strings have been replaced with string versions.